Rearrange proportional amounts helpers
