+++ src/entities/utils/index.ts
@@ -1,10 +1,12 @@
-export * from './calculateProportionalAmounts';
 export * from './cowAmmHelpers';
 export * from './doAddLiquidityQuery';
 export * from './getAmounts';
+export * from './getPoolStateWithBalancesV2';
+export * from './getPoolStateWithBalancesV3';
 export * from './getSortedTokens';
 export * from './getValue';
 export * from './parseAddLiquidityArgs';
 export * from './parseInitializeArgs';
+export * from './proportionalAmountsHelpers';
 export * from './replaceWrapped';
 export * from './validateNestedPoolState';
